rename slice to reducer
