+++ client/src/components/AlcoCounter/alcoComponents/Calendar/Calendar.tsx
@@ -18,7 +18,7 @@ import {
   useAppSelector,
   useAppDispatch,
 } from "../../../../store/hooks/redux";
-import { alcoActions } from "../../../../store/reducer/alcoSlice";
+import { alcoActions } from "../../../../store/reducer/alcoReducer";
 import { getAlcoYear } from "../../../../store/reducer/http/alcoActions";
 import { DayInfo } from "../../../../types/alcoTypes";
 import updateLocale from "dayjs/plugin/updateLocale";
