+++ client/src/components/TopMenu/SelectTheme.tsx
@@ -9,7 +9,7 @@ import {
   useAppDispatch,
   useAppSelector,
 } from "../../store/hooks/redux";
-import { userActions } from "../../store/reducer/userSlice";
+import { userActions } from "../../store/reducer/userReducer";
 import { UserThemes } from "../../types/userTypes";
 
 export function SelectTheme() {
