+++ client/src/components/TopMenu/UserAuthBar/Logout.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from "react";
 import { Stack } from "@mui/material";
 import { LoadingButton } from "@mui/lab";
-import { userActions } from "../../../store/reducer/userSlice";
+import { userActions } from "../../../store/reducer/userReducer";
 import { useDispatch } from "react-redux";
 import { ModalOpen } from "../../../types/userTypes";
