+++ client/src/components/TopMenu/UserAuthBar/UserAuthBar.tsx
@@ -9,7 +9,7 @@ import {
   useAppSelector,
   useAppDispatch,
 } from "../../../store/hooks/redux";
-import { userActions } from "../../../store/reducer/userSlice";
+import { userActions } from "../../../store/reducer/userReducer";
 import { Message } from "../../ui";
 
 export default function UserAuthBar() {
