+++ client/src/store/index.tsx
@@ -4,10 +4,10 @@ import {
   Action,
   combineReducers,
 } from "@reduxjs/toolkit";
-import alcoReducer from "./reducer/alcoSlice";
-import salaryReducer from "./reducer/salarySlice";
-import userReducer from "./reducer/userSlice";
-import serviceReducer from "./reducer/serviceSlice";
+import alcoReducer from "./reducer/alcoReducer";
+import salaryReducer from "./reducer/salaryReducer";
+import userReducer from "./reducer/userReducer";
+import serviceReducer from "./reducer/serviceReducer";
 
 const rootReducer = combineReducers({
   userReducer,
