+++ client/src/store/reducer/salaryReducer.ts
